+++ lib/src/server/handler.dart
@@ -20,7 +20,7 @@ import 'package:http2/transport.dart';
 
 import '../shared/codec.dart';
 import '../shared/codec_registry.dart';
-import '../shared/io_bits/io_bits.dart' show X509Certificate;
+import '../shared/io_bits/io_bits.dart' show InternetAddress, X509Certificate;
 import '../shared/message.dart';
 import '../shared/status.dart';
 import '../shared/streams.dart';
@@ -63,7 +63,9 @@ class ServerHandler extends ServiceCall {
   bool _isCanceled = false;
   bool _isTimedOut = false;
   Timer? _timeoutTimer;
+
   final X509Certificate? _clientCertificate;
+  final InternetAddress? _remoteAddress;
 
   ServerHandler({
     required ServerTransportStream stream,
@@ -71,12 +73,14 @@ class ServerHandler extends ServiceCall {
     required List<Interceptor> interceptors,
     required CodecRegistry? codecRegistry,
     X509Certificate? clientCertificate,
+    InternetAddress? remoteAddress,
     GrpcErrorHandler? errorHandler,
   })  : _stream = stream,
         _serviceLookup = serviceLookup,
         _interceptors = interceptors,
         _codecRegistry = codecRegistry,
         _clientCertificate = clientCertificate,
+        _remoteAddress = remoteAddress,
         _errorHandler = errorHandler;
 
   @override
@@ -100,14 +104,21 @@ class ServerHandler extends ServiceCall {
   @override
   X509Certificate? get clientCertificate => _clientCertificate;
 
+  @override
+  InternetAddress? get remoteAddress => _remoteAddress;
+
   void handle() {
     _stream.onTerminated = (_) => cancel();
 
     _incomingSubscription = _stream.incomingMessages
         .transform(GrpcHttpDecoder())
         .transform(grpcDecompressor(codecRegistry: _codecRegistry))
-        .listen(_onDataIdle,
-            onError: _onError, onDone: _onDoneError, cancelOnError: true);
+        .listen(
+          _onDataIdle,
+          onError: _onError,
+          onDone: _onDoneError,
+          cancelOnError: true,
+        );
     _stream.outgoingMessages.done.then((_) {
       cancel();
     });
@@ -342,8 +353,11 @@ class ServerHandler extends ServiceCall {
   }
 
   @override
-  void sendTrailers(
-      {int? status = 0, String? message, Map<String, String>? errorTrailers}) {
+  void sendTrailers({
+    int? status = 0,
+    String? message,
+    Map<String, String>? errorTrailers,
+  }) {
     _timeoutTimer?.cancel();
 
     final outgoingTrailersMap = <String, String>{};